Add EndToEndEncryptionTest to test files encrypted on disk
This test creates a database with a key using the DatastoreManager.
It then looks at the JSON database, the indexes database and the
attachment file created to check they are encrypted or not (based
on the `test.sqlcipher.passphrase` define.

This isn't ideal, as the encryption code path will only be tested
when `test.sqlcipher.passphrase` is defined, which it isn't by
default. This is hard to parameterise until we support SQLCipher
on Java SE -- obviously the ideal would be to always run one
set of the tests where encryption is on, one where it's off.